Add files via upload
The Indian Premier League (IPL) is a professional Twenty20 cricket league in India, established in 2008 by the Board of Control for Cricket in India (BCCI). This analysis includes the Machine Learning analysis of the winning probability of  Royal Challengers Bangalore, Deep Learning analysis in predicting the winners using TensorFlow, most run-scored balls in an over, the most valuable player in the IPL why Royal Challengers are failing in winning the title, the best Coach etc.